+++ FA2sp/Miscs/MultiSelection.cpp
@@ -4,9 +4,13 @@
 #include <CIsoView.h>
 #include <Helpers/Macro.h>
 #include <Drawing.h>
+#include <CTileTypeClass.h>
 
 #include "../Ext/CIsoView/Body.h"
 #include "../Ext/CMapData/Body.h"
+
+#include <algorithm>
+#include <span>
 
 // #include "../Source/CIsoView.h"
 
@@ -72,6 +76,118 @@ inline bool MultiSelection::IsSelected(int X, int Y)
     return SelectedCoords.find(MapCoord{ X,Y }) != SelectedCoords.end();
 }
 
+void MultiSelection::Copy()
+{
+    std::vector<MyClipboardData> datas;
+    for (const auto& coord : SelectedCoords)
+    {
+        auto pCell = CMapData::Instance->GetCellAt(coord.X, coord.Y);
+        MyClipboardData data;
+        data.X = coord.X;
+        data.Y = coord.Y;
+        data.Overlay = pCell->Overlay;
+        data.OverlayData = pCell->OverlayData;
+        data.TileIndex = pCell->TileIndex;
+        data.TileIndexHiPart = pCell->TileIndexHiPart;
+        data.TileSubIndex = pCell->TileSubIndex;
+        data.Height = pCell->Height;
+        data.IceGrowth = pCell->IceGrowth;
+        data.Flag = pCell->Flag;
+        datas.push_back(data);
+    }
+
+    auto hGlobal = GlobalAlloc(GMEM_SHARE | GMEM_MOVEABLE, 8 + sizeof(MyClipboardData) * datas.size());
+    if (hGlobal == NULL)
+    {
+        MessageBox(NULL, "Error", "Failed to allocate global memory!", MB_OK);
+        return;
+    }
+
+    auto pBuffer = GlobalLock(hGlobal);
+    if (pBuffer == nullptr)
+    {
+        MessageBox(NULL, "Error", "Failed to lock hGlobal handle!", MB_OK);
+        return;
+    }
+    while (GlobalUnlock(hGlobal))
+        ;
+
+    reinterpret_cast<int*>(pBuffer)[0] = 0; // Flag indicate this is multiselection
+    reinterpret_cast<size_t*>(pBuffer)[1] = datas.size();
+    memcpy(reinterpret_cast<char*>(pBuffer) + 8, datas.data(), sizeof(MyClipboardData) * datas.size());
+
+    OpenClipboard(CFinalSunApp::Instance->m_pMainWnd->m_hWnd);
+    EmptyClipboard();
+    if (FALSE == SetClipboardData(CFinalSunApp::Instance->ClipboardFormat, hGlobal))
+        MessageBox(NULL, "Failed to set clipboard data", "Error", 0);
+    CloseClipboard();
+}
+
+void MultiSelection::Paste(int X, int Y, int nBaseHeight, MyClipboardData* data, size_t length)
+{
+    std::span<MyClipboardData> cells {data, data + length};
+    
+    RECT rect
+    { 
+        std::numeric_limits<LONG>::max(), 
+        std::numeric_limits<LONG>::max(), 
+        std::numeric_limits<LONG>::min(),
+        std::numeric_limits<LONG>::min() 
+    };
+    for (const auto& cell : cells)
+    {
+        if (cell.X < rect.left)
+            rect.left = cell.X;
+        if (cell.X > rect.right)
+            rect.right = cell.X;
+        if (cell.Y < rect.top)
+            rect.top = cell.Y;
+        if (cell.Y > rect.bottom)
+            rect.bottom = cell.Y;
+    }
+
+    const MapCoord center = { (rect.left + rect.right) / 2, (rect.top + rect.bottom) / 2 };
+
+    auto lowest_height = std::numeric_limits<unsigned char>::min();
+    for (const auto& cell : cells)
+    {
+        int offx = cell.X - center.X;
+        int offy = cell.Y - center.Y;
+
+        const auto pCell = CMapData::Instance->TryGetCellAt(X + offx, Y + offy);
+        if (pCell->Height < lowest_height)
+            lowest_height = pCell->Height;
+    }
+
+    nBaseHeight += lowest_height;
+    for (const auto& cell : cells)
+    {
+        int offx = cell.X - center.X;
+        int offy = cell.Y - center.Y;
+
+        auto nCellIndex = CMapData::Instance->GetCoordIndex(X + offx, Y + offy);
+        if (nCellIndex < 0 || nCellIndex >= CMapData::Instance->CellDataCount)
+            continue;
+
+        auto pCell = CMapData::Instance->GetCellAt(nCellIndex);
+        
+        CMapData::Instance->DeleteTiberium(pCell->Overlay, pCell->OverlayData);
+        pCell->Overlay = cell.Overlay;
+        pCell->OverlayData = cell.OverlayData;
+        CMapData::Instance->AddTiberium(pCell->Overlay, pCell->OverlayData);
+        
+        pCell->TileIndex = cell.TileIndex;
+        pCell->TileIndexHiPart = cell.TileIndexHiPart;
+        pCell->TileSubIndex = cell.TileSubIndex;
+        
+        pCell->Height = std::clamp(cell.Height + nBaseHeight, 0, 14);
+
+        pCell->IceGrowth = cell.IceGrowth;
+        pCell->Flag = cell.Flag;
+
+        CMapData::Instance->UpdateMapPreviewAt(X + offx, Y + offy);
+    }
+}
 
 DEFINE_HOOK(456EFC, CIsoView_OnMouseMove_MultiSelect_ReverseStatus, 6)
 {
@@ -306,32 +422,58 @@ DEFINE_HOOK(433F70, CFinalSunDlg_Tools_HideSingleField, 5)
     return 0x433F83;
 }
 
-//DEFINE_HOOK(435F10, CFinalSunDlg_Tools_Copy, 7)
-//{
-//    GET(CFinalSunDlg*, pThis, ECX);
-//
-//    pThis->PlaySound(CFinalSunDlg::FASoundType::Normal);
-//
-//    if (MultiSelection::GetCount())
-//    {
-//
-//    }
-//    else
-//        CIsoView::CurrentCommand->Command = FACurrentCommand::TileCopy;
-//
-//    return 0x435F24;
-//}
-//
-//DEFINE_HOOK(4C3850, CMapData_PasteAt, 8)
-//{
-//    GET_STACK(const int, X, 0x4);
-//    GET_STACK(const int, Y, 0x8);
-//    GET_STACK(const char, nBaseHeight, 0xC);
-//
-//
-//
-//    return 0x4C388B;
-//}
+DEFINE_HOOK(435F10, CFinalSunDlg_Tools_Copy, 7)
+{
+    if (!ExtConfigs::EnableMultiSelection)
+        return 0;
+
+    GET(CFinalSunDlg*, pThis, ECX);
+
+    pThis->PlaySound(CFinalSunDlg::FASoundType::Normal);
+
+    if (MultiSelection::GetCount())
+        MultiSelection::Copy();
+    else
+        CIsoView::CurrentCommand->Command = FACurrentCommand::TileCopy;
+
+    return 0x435F24;
+}
+
+DEFINE_HOOK(4C3850, CMapData_PasteAt, 8)
+{
+    if (!ExtConfigs::EnableMultiSelection)
+        return 0;
+
+    GET_STACK(const int, X, 0x4);
+    GET_STACK(const int, Y, 0x8);
+    GET_STACK(const char, nBaseHeight, 0xC);
+
+    OpenClipboard(CFinalSunApp::Instance->m_pMainWnd->m_hWnd);
+    HANDLE hData = GetClipboardData(CFinalSunApp::Instance->ClipboardFormat);
+    auto ptr = GlobalLock(hData);
+    
+    if (ptr)
+    {
+        if (reinterpret_cast<int*>(ptr)[0] == 0) // Multiselection
+        {
+            const auto length = reinterpret_cast<size_t*>(ptr)[1];
+            const auto p = reinterpret_cast<MultiSelection::MyClipboardData*>(reinterpret_cast<char*>(ptr) + 8);
+            MultiSelection::Paste(X, Y, nBaseHeight, p, length);
+            GlobalUnlock(hData);
+            CloseClipboard();
+            return 0x4C388B;
+        }
+        else // Default selection
+        {
+            GlobalUnlock(hData);
+            CloseClipboard();
+            return 0;
+        }
+    }
+
+    CloseClipboard();
+    return 0x4C388B;
+}
 
 DEFINE_HOOK(474FE0, CIsoView_Draw_MultiSelectionMoney, 5)
 {